+++ src/ansys/fluent/visualization/__init__.py
@@ -24,5 +24,5 @@ def version_info() -> str:
 
 
 from ansys.fluent.visualization._config import get_config, set_config  # noqa: F401
-from ansys.fluent.visualization.graphic import Graphics  # noqa: F401
+from ansys.fluent.visualization.graphics import Graphics  # noqa: F401
 from ansys.fluent.visualization.plotter import Plots  # noqa: F401
